add editing of records

+++ app.js
@@ -37,29 +37,54 @@ app.get('', urlencodedParser, (request, response) => {
     get_all(response);    
 });
 
-app.post("/delete", urlencodedParser, function (request, response) {
+app.post("/update", urlencodedParser, function (request, response) {
     if(!request.body) return response.sendStatus(400);
 
-    connection.query('DELETE FROM Spending WHERE spending_id=' + request.body.spending_id 
-                        + ' AND user_id=1', function (err, result) {
-        if (err) throw err;
-        });
+    if (typeof(request.body.delete) != "undefined" && request.body.delete !== null) {
+        connection.query('DELETE FROM Spending WHERE spending_id=' + request.body.spending_id 
+                            + ' AND user_id=1', function (err, result) {
+            if (err) throw err;
+            });
 
-    get_all(response); 
+        get_all(response); 
+    } else if (typeof(request.body.edit) != "undefined" && request.body.edit !== null) {
+
+        connection.query('SELECT * FROM Spending WHERE spending_id=' + request.body.spending_id + ' AND user_id=1', function (err, currFin) {
+            if (err) throw err;
+            connection.query('SELECT * FROM Spending ORDER BY ' + sorting_info[0] + (sorting_info[1] == false ? " DESC":""), function (err, result) {
+                if (err) throw err;
+                console.log(currFin);
+                response.render('index', { fins: result, title: 'Трекер финансов', sorting_info:sorting_info, submitValue: "Обновить", editFin: currFin[0]});
+            });
+        });
+        
+    }
  });
 
 app.post("/add", urlencodedParser, function (request, response) {
     if(!request.body) return response.sendStatus(400);
 
-    connection.query('INSERT Spending(user_id, sum, date, category, description, income) VALUES (?,?,?,?,?,?)',
-    [
-    1,
-    request.body.sum * 100,
-    request.body.date,
-    request.body.category,
-    request.body.description,
-    request.body.type == 'income'
-    ]);
+    if (request.body.edit_id == -1) {
+        connection.query('INSERT Spending(user_id, sum, date, category, description, income) VALUES (?,?,?,?,?,?)',
+        [
+        1,
+        request.body.sum * 100,
+        request.body.date,
+        request.body.category,
+        request.body.description,
+        request.body.type == 'income'
+        ]); 
+    } else {
+        connection.query('UPDATE Spending SET sum = ?, date = ?, category = ?, description = ?, income = ? WHERE spending_id = ?  AND user_id=1',
+        [
+        request.body.sum * 100,
+        request.body.date,
+        request.body.category,
+        request.body.description,
+        request.body.type == 'income',
+        request.body.edit_id
+        ]); 
+    }
 
     get_all(response); 
  });
@@ -71,6 +96,7 @@ app.post("/sort", urlencodedParser, function (request, response) {
 
     get_all(response); 
  });
+
 
 app.listen(port, () => console.log('App listening...')); 
 
@@ -90,6 +116,6 @@ function define_sort_type(sorting_type, sorting_direction)
 function get_all(response) {
     connection.query('SELECT * FROM Spending ORDER BY ' + sorting_info[0] + (sorting_info[1] == false ? " DESC":""), function (err, result) {
         if (err) throw err;
-        response.render('index', { fins: result, title: 'Трекер финансов', sorting_info:sorting_info});
+        response.render('index', { fins: result, title: 'Трекер финансов', sorting_info:sorting_info, submitValue: "Добавить" });
     });
 }
